Fix sufficient S/MIME condition question
This is a tricky one, imo similarily to the necessary
condition question (1009), you need to somehow get the
CA certificate from the trusted party, otherwise the
attacker could send you his own certificate and sign the
document with the key defined in it.

I suppose that if he want a sufficient condition then he
wants only one option to be selected. Then I guess that having
access to some CA center and downloading the certificates of CA
and the senders from it would be enough to verify the signature.